pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptbeacon"
version = "0.1.0"
description = "LLM visibility monitoring for brands - track how your brand appears in AI-generated responses"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yotam Braun" }
]
maintainers = [
    { name = "Yotam Braun" }
]
keywords = [
    "llm",
    "brand",
    "visibility",
    "monitoring",
    "ai",
    "chatgpt",
    "claude",
    "gemini",
    "brand-monitoring",
    "seo",
    "analytics",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "litellm>=1.0.0,<2.0.0",
    "httpx>=0.26.0,<1.0.0",
    "instructor>=0.4.0,<2.0.0",
    "duckdb>=0.10.0,<2.0.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
cli = [
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.31.0",
]
pandas = [
    "pandas>=2.0.0",
]
all = [
    "promptbeacon[cli,pandas]",
]

[project.scripts]
promptbeacon = "promptbeacon.cli.main:app"

[project.urls]
Homepage = "https://github.com/yotambraun/promptbeacon"
Documentation = "https://github.com/yotambraun/promptbeacon/tree/main/docs"
Repository = "https://github.com/yotambraun/promptbeacon"
Changelog = "https://github.com/yotambraun/promptbeacon/blob/main/CHANGELOG.md"
Issues = "https://github.com/yotambraun/promptbeacon/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/promptbeacon"]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B905",   # zip strict
]

[tool.ruff.lint.isort]
known-first-party = ["promptbeacon"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/promptbeacon"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
